fix(map legends): Fixes
- Fixed issues with map legends not being removed when corresponding layer is removed
