+++ src/components/Map/components/MapArea/interfaces/index.ts
@@ -1,5 +1,6 @@
 import { ControlPosition } from "leaflet";
 import type { MapContainerProps } from "react-leaflet";
+import { CustomBoundaryLayer, CustomThematicPrimitiveLayer } from "../../MapLayer/interfaces";
 
 export interface MapControls {
   position: ControlPosition;
@@ -21,5 +22,6 @@ export interface MapAreaProps {
   controls?: MapControls[];
   mapOptions?: MapContainerProps;
   legends?: MapLegendConfig;
+  layers: Array<CustomThematicPrimitiveLayer | CustomBoundaryLayer>;
   key?: string;
 }
