+++ src/components/Map/components/MapProvider/hooks/index.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { MapOrgUnitContext, MapPeriodContext } from "../../../state";
+import { MapLayersContext, MapOrgUnitContext, MapPeriodContext } from "../../../state";
 
 export function useMapOrganisationUnit() {
   return useContext(MapOrgUnitContext);
@@ -7,4 +7,8 @@ export function useMapOrganisationUnit() {
 
 export function useMapPeriods() {
   return useContext(MapPeriodContext);
+}
+
+export function useMapLayers() {
+  return useContext(MapLayersContext);
 }
