+++ src/components/Map/state/index.tsx
@@ -1,6 +1,6 @@
 import type { OrgUnitSelection } from "@hisptz/dhis2-utils";
 import { createContext } from "react";
-import { MapLayer } from "../components/MapLayer/interfaces";
+import { CustomMapLayer } from "../components/MapLayer/interfaces";
 import { MapOrgUnit } from "../interfaces";
 
 export const MapOrgUnitContext = createContext<{
@@ -21,7 +21,9 @@ export const MapPeriodContext = createContext<
 });
 
 export const MapLayersContext = createContext<{
-  layers: MapLayer[];
+  layers: CustomMapLayer[];
+  updateLayer: (id: string, updatedLayer: CustomMapLayer) => void;
 }>({
   layers: [],
+  updateLayer: () => {},
 });
